+++ README.md
@@ -1,35 +1,9 @@
-# Stanford Core NLP Service
+# Stanford NLP Service
 
-The Stanford Core NLP service exposes several annotation pipelines for use by external services. They are:
+The Stanford Core NLP service processes documents received from the RabbitMQ server.
 
-1. *Segmentation* provides `Token` and `Sentence` annotations.
-1. *Part of Speech Tagging* adds *pos* tags to tokens.
-1. *Lemmatization* adds *lemma* tags to the tokens.
-1. *Named Entity Recognition* provides `NamedEntity` annotations. 
+**NOTE** The Stanford NLP service is not (yet) used for NLP processing in the web application.
 
-Each of the pipelines executes all the annotators from the previous pipelines, that is, the named entity recognizer also produces token and sentence annotions with lemmas and part of speech tags.
-  
+# TODO
 
-## Connection Details
-
-**Host** *rabbitmq.lappsgrid.org/nlp*<br/>
-**Exchange** *stanford*<br/>
-**Mailbox** *pipelines*
-
-| Property | Value |
-|------|-----------------------|
-| Host | rabbitmq.lappsgrid.org|
-| Exchange | askme |
-| Mailbox | nlp |
-
-
-```
-PostOffice po = new PostOffice("askme", "rabbitmq.lappsgrid.org");
-Message message = new Message()
-...
-message.route("nlp")
-po.send(message)
-```
-
-## Messages
-
+Write more documentation.
