+++ pom.xml
@@ -3,9 +3,9 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>groovy-parent-pom</artifactId>
-        <groupId>org.lappsgrid.maven</groupId>
-        <version>1.1.2</version>
+        <artifactId>parent-pom</artifactId>
+        <groupId>org.lappsgrid.eager.mining</groupId>
+        <version>1.0.0</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
@@ -14,11 +14,9 @@
     <version>1.0.0</version>
 
     <properties>
-        <stanford.version>3.9.2</stanford.version>
-        <metrics.version>3.2.6</metrics.version>
+        <stanford.version>3.9.1</stanford.version>
     </properties>
     <build>
-        <finalName>service</finalName>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -31,33 +29,16 @@
                             <goal>single</goal>
                         </goals>
                         <configuration>
+                            <finalName>stanford</finalName>
                             <appendAssemblyId>false</appendAssemblyId>
                             <descriptorRefs>
                                 <descriptorRef>jar-with-dependencies</descriptorRef>
                             </descriptorRefs>
                             <archive>
                                 <manifest>
-                                    <mainClass>org.lappsgrid.askme.nlp.Main</mainClass>
+                                    <mainClass>org.lappsgrid.eager.mining.web.nlp.stanford.Main</mainClass>
                                 </manifest>
                             </archive>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.anc.maven.plugins</groupId>
-                <artifactId>anc-maven-plugin</artifactId>
-                <version>1.1.0</version>
-                <executions>
-                    <execution>
-                        <id>version-file</id>
-                        <phase>generate-resources</phase>
-                        <goals>
-                            <goal>version</goal>
-                            <goal>build</goal>
-                        </goals>
-                        <configuration>
-                            <buildFile>src/main/resources/build</buildFile>
                         </configuration>
                     </execution>
                 </executions>
@@ -66,14 +47,14 @@
     </build>
     <dependencies>
         <dependency>
-            <groupId>org.lappsgrid</groupId>
+            <groupId>org.lappsgrid.eager.mining</groupId>
             <artifactId>rabbitmq</artifactId>
-            <version>1.0.0</version>
+            <version>1.2.0</version>
         </dependency>
         <dependency>
-            <groupId>org.lappsgrid.askme</groupId>
+            <groupId>org.lappsgrid.eager.mining</groupId>
             <artifactId>core</artifactId>
-            <version>1.0.0-SNAPSHOT</version>
+            <version>1.2.0</version>
         </dependency>
         <dependency>
             <groupId>edu.stanford.nlp</groupId>
@@ -89,33 +70,55 @@
         <dependency>
             <groupId>org.lappsgrid</groupId>
             <artifactId>serialization</artifactId>
-            <version>${lapps.serialization.version}</version>
+            <version>2.6.0</version>
         </dependency>
         <dependency>
             <groupId>org.lappsgrid</groupId>
             <artifactId>discriminator</artifactId>
-            <version>${lapps.discriminator.version}</version>
+            <version>2.3.3</version>
         </dependency>
         <dependency>
             <groupId>org.lappsgrid</groupId>
             <artifactId>vocabulary</artifactId>
-            <version>${lapps.vocabulary.version}</version>
+            <version>2.4.1</version>
+        </dependency>
+        <!--
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-api</artifactId>
+            <version>2.11.0</version>
         </dependency>
         <dependency>
-            <groupId>io.dropwizard.metrics</groupId>
-            <artifactId>metrics-core</artifactId>
-            <version>${metrics.version}</version>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-core</artifactId>
+            <version>2.11.0</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-slf4j-impl</artifactId>
+            <version>2.11.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>1.7.25</version>
+        </dependency>
+        <dependency>
+            <groupId>ch.qos.logback</groupId>
+            <artifactId>logback-classic</artifactId>
+            <version>1.2.3</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-to-slf4j</artifactId>
+            <version>2.11.0</version>
+        </dependency>
+        -->
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>4.12</version>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-            <type>jar</type>
         </dependency>
     </dependencies>
