--- src/main/groovy/org/lappsgrid/askme/nlp/Main.groovy
@@ -1,235 +1,0 @@
-package org.lappsgrid.askme.nlp
-
-import com.codahale.metrics.Meter
-import com.codahale.metrics.Timer
-import groovy.transform.TypeChecked
-import groovy.util.logging.Slf4j
-import org.lappsgrid.askme.core.Configuration
-
-//import org.lappsgrid.eager.mining.core.jmx.Registry
-import org.lappsgrid.rabbitmq.Message
-import org.lappsgrid.rabbitmq.RabbitMQ
-import org.lappsgrid.rabbitmq.topic.MailBox
-import org.lappsgrid.rabbitmq.topic.PostOffice
-
-//import org.lappsgrid.eager.mining.core.Configuration
-//import org.lappsgrid.eager.mining.core.jmx.Registry
-//import org.lappsgrid.eager.rabbitmq.Message
-//import org.lappsgrid.eager.rabbitmq.topic.MailBox
-//import org.lappsgrid.eager.rabbitmq.topic.PostOffice
-import org.lappsgrid.serialization.Serializer
-
-import java.util.concurrent.BlockingQueue
-import java.util.concurrent.LinkedBlockingQueue
-import java.util.concurrent.ThreadPoolExecutor
-import java.util.concurrent.TimeUnit
-
-/**
- *
- */
-@TypeChecked
-@Slf4j('logger')
-class Main implements MainMBean {
-
-    static final Configuration config = new Configuration()
-    static {
-        System.setProperty(RabbitMQ.USERNAME_PROPERTY, config.USERNAME)
-        System.setProperty(RabbitMQ.PASSWORD_PROPERTY, config.PASSWORD)
-    }
-
-    static final String MAILBOX = "nlp"
-
-    static final String SEGMENTER = "segmenter"
-    static final String POS = "pos"
-    static final String LEMMAS = "lemmas"
-    static final String NER = "ner"
-
-    /** The number of documents processed. */
-//    final Meter count = Registry.meter('nlp', 'count')
-    /** The number of errors encountered. */
-//    final Meter errors = Registry.meter('nlp', 'errors')
-    /** Processing time for documents. */
-//    final Timer timer = Registry.timer('nlp', 'timer')
-
-    // Since we explicitly create a ThreadPoolExecutor we also need to
-    // explicitly create the BlockingQueue used by the pool.
-    private BlockingQueue<Runnable> queue
-    ThreadPoolExecutor pool
-
-    /** The Stanford service pipelines to be executed. */
-    Map<String,Pipeline> pipelines
-
-    /** Object used to block/wait until the queue is closed. */
-    Object semaphore
-
-    /** Where outgoing messages are sent. */
-    PostOffice post
-
-    /** Where we receive incoming messages. */
-    MailBox box
-
-    Main() {
-        // Initialize the pipelines.  Stanford Core NLP claims to be
-        // thread-safe so we will take them at their word.
-        logger.info("Rabbit username: {}", System.getProperty(RabbitMQ.USERNAME_PROPERTY))
-        pipelines = [
-            segmenter: Pipeline.Segmenter(),
-            pos: Pipeline.Tagger(),
-            lemmas: Pipeline.Lemmatizer(),
-            ner: Pipeline.NamedEntityRecognizer()
-        ]
-
-        this.semaphore = new Object()
-        this.post = new PostOffice(config.EXCHANGE, config.HOST)
-        logger.trace("Rabbit host: {}", config.HOST)
-        logger.trace("Rabbit Exchange: {}", config.EXCHANGE)
-        // Configure our thread pool executor
-        queue = new LinkedBlockingQueue<>()
-        int minCores = 2
-        int maxCores = 2
-        int totalCores = Runtime.getRuntime().availableProcessors()
-        if (totalCores <= 2) {
-            minCores = maxCores = totalCores
-        }
-        else {
-            maxCores = (int)(totalCores / 2)
-//            minCores = maxCores // 2
-        }
-
-        pool = new ThreadPoolExecutor(minCores, maxCores, 30, TimeUnit.SECONDS, queue)
-        logger.debug("Threadpool size {}-{}", minCores, maxCores)
-
-    }
-
-    void start() {
-
-        logger.info("Staring Stanford NLP service.")
-        box = new MailBox(config.EXCHANGE, MAILBOX, config.HOST) {
-            @Override
-            void recv(String json) {
-                Message message
-                try {
-                    logger.debug("Received message. Size: {}", json.length())
-                    message = Serializer.parse(json, Message)
-                }
-                catch (Exception e) {
-                    error(e.getMessage())
-                    return
-                }
-
-                if ('EXIT' == message.command || 'QUIT' == message.command) {
-                    logger.info("Received a shutdown message.")
-                    stop()
-                    return
-                }
-                if(message.command == 'PING') {
-                    logger.info('Received PING message from and sending response to {}', message.route[0])
-                    Message response = new Message()
-                    response.setBody('PONG')
-                    response.setCommand('PONG')
-                    response.setRoute(message.route)
-                    Main.this.post.send(response)
-                    return
-                }
-                if (message.route.size() == 0) {
-                    // There is nowhere to send the result so we have nothing to do.
-                    error("NLP tools were sent data but have no route defined.")
-                    return
-                }
-
-                logger.debug("Looking up the pipeline for {}", message.command)
-                Pipeline pipeline = Main.this.pipelines[message.command]
-                if (pipeline == null) {
-                    error("Invalid pipeline " + message.command)
-                    return
-                }
-
-                logger.debug("Staring a worker.")
-                //Worker worker = new Worker(pipeline, message, Main.this.post, Main.this.timer, Main.this.count, Main.this.errors)
-                Worker worker = new Worker(pipeline, message, Main.this.post, null, null, null)
-                Main.this.pool.execute(worker)
-            }
-        }
-    }
-
-    String stats() {
-        return String.format("[monitor] [%d/%d/%d] Active: %d\nCompleted: %d\nTask: %d\nisShutdown: %s\nisTerminated: %s",
-                this.pool.getPoolSize(),
-                this.pool.getCorePoolSize(),
-                this.pool.getMaximumPoolSize(),
-                this.pool.getActiveCount(),
-                this.pool.getCompletedTaskCount(),
-                this.pool.getTaskCount(),
-                this.pool.isShutdown(),
-                this.pool.isTerminated())
-    }
-
-    String increase() {
-        int cores = Runtime.getRuntime().availableProcessors()
-        int size = pool.maximumPoolSize + 1
-        pool.maximumPoolSize = size
-        if (size < cores) {
-            return "OK size is now " + size
-        }
-        return "Pool size is larger than available processors"
-    }
-
-    String decrease() {
-        if (pool.maximumPoolSize > 1) {
-            pool.maximumPoolSize = pool.maximumPoolSize - 1
-            return "OK"
-        }
-        return "Already at one"
-    }
-
-    void stop() {
-        logger.info("Stopping the Stanford NLP service")
-        pool.shutdown()
-        if (!pool.awaitTermination(30, TimeUnit.SECONDS)) {
-            logger.warn("Thread pool failed to shutdown cleanly.  Trying to force a shutdown.")
-            pool.shutdownNow()
-        }
-
-        logger.trace("Closing post office")
-//        close(post)
-        post.close()
-        logger.trace("Closing mailbox")
-//        close(box)
-        box.close()
-        logger.trace("Notifying the semaphore")
-        synchronized (semaphore) {
-            semaphore.notifyAll()
-        }
-        logger.trace("The semaphore has been notified")
-    }
-
-//    void close(def thing) {
-//        if (thing == null) {
-//            return
-//        }
-//        try {
-//            thing.close()
-//        }
-//        catch (Exception e) {
-//        }
-//    }
-
-    private void error(String message) {
-        logger.error(message)
-//        errors.mark()
-    }
-
-    static void main(String[] args) {
-        Main app = new Main()
-//        Registry.register(app, "org.lappsgrid.eager.mining.nlp.stanford.Main:type=Main")
-//        Registry.startJmxReporter()
-        app.start()
-
-        // Wait until another thread calls notify() or notifyAll() on the semaphore.
-        synchronized (app.semaphore) {
-            app.semaphore.wait()
-        }
-        logger.info("Stanford NLP service terminated.")
-        System.exit(0)
-    }
-}
