+++ src/main/groovy/org/lappsgrid/eager/mining/web/nlp/stanford/Pipeline.groovy
@@ -1,17 +1,17 @@
-package org.lappsgrid.askme.nlp
+package org.lappsgrid.eager.mining.web.nlp.stanford
 
 import edu.stanford.nlp.ling.CoreLabel
 import edu.stanford.nlp.pipeline.CoreDocument
-import edu.stanford.nlp.pipeline.CoreEntityMention
 import edu.stanford.nlp.pipeline.CoreSentence
 import edu.stanford.nlp.pipeline.StanfordCoreNLP
 import edu.stanford.nlp.util.Pair
-import groovy.transform.TypeChecked
 import groovy.util.logging.Slf4j
 import org.lappsgrid.serialization.LifException
 import org.lappsgrid.serialization.lif.Annotation
 import org.lappsgrid.serialization.lif.Container
 import org.lappsgrid.serialization.lif.View
+//import org.slf4j.Logger
+//import org.slf4j.LoggerFactory
 
 import static org.lappsgrid.discriminator.Discriminators.*;
 import org.lappsgrid.vocabulary.Features
@@ -19,30 +19,14 @@ import org.lappsgrid.vocabulary.Features
 /**
  *
  */
-@TypeChecked
 @Slf4j('logger')
 public class Pipeline
 {
-    // Factory methods
+    StanfordCoreNLP pipeline;
 
-    static public Pipeline Segmenter() {
-        return new Pipeline("tokenize,ssplit")
-    }
-    static public Pipeline Tagger() {
-        return new Pipeline("tokenize,ssplit,pos")
-    }
-    static public Pipeline Lemmatizer() {
-        return new Pipeline("tokenize,ssplit,pos,lemma")
-    }
-    static public Pipeline NamedEntityRecognizer() {
-        return new Pipeline("tokenize,ssplit,pos,lemma,ner")
-    }
-
-    private StanfordCoreNLP pipeline;
-
-    protected Pipeline(String annotators) {
+    public  Pipeline() {
         Properties props = new Properties();
-        props.setProperty("annotators", annotators);
+        props.setProperty("annotators", "tokenize,ssplit,pos,lemma");
         pipeline = new StanfordCoreNLP(props);
     }
 
@@ -76,27 +60,20 @@ public class Pipeline
         logger.trace('Processing tokens')
         View tokens = container.newView();
         id = 0;
-        int pos = 0;
-        int lemma = 0;
         for (CoreLabel token : document.tokens()) {
             int start = token.beginPosition();
             int end = token.endPosition();
             Annotation a = tokens.newAnnotation("tok-" + (id++), Uri.TOKEN, start, end);
-            if (set(a, Features.Token.LEMMA, token.lemma())) ++lemma
-            if (set(a, Features.Token.PART_OF_SPEECH, token.tag())) ++ pos
+            set(a, Features.Token.LEMMA, token.lemma());
+            set(a, Features.Token.PART_OF_SPEECH, token.tag());
             set(a, Features.Token.WORD, token.word());
             set(a, "category", token.category());
         }
         if (id > 0) {
             tokens.addContains(Uri.TOKEN, this.getClass().getName(), "stanford");
         }
-        if (pos > 0) {
-            tokens.addContains(Uri.POS, this.class.name, "stanford")
-        }
-        if (lemma > 0) {
-            tokens.addContains(Uri.LEMMA, this.class.name, 'stanford')
-        }
 
+        /*
         View ner = container.newView();
         id = 0;
         for (CoreEntityMention mention : document.entityMentions()) {
@@ -107,19 +84,16 @@ public class Pipeline
         if (id > 0) {
             ner.addContains(Uri.NE, this.getClass().getName(), "stanford");
         }
-        else {
-            container.views.remove(ner)
-        }
+        */
 //        return new Data(Uri.LIF, container).asPrettyJson();
         logger.debug('Processing complete')
         return container
     }
 
-    protected boolean set(Annotation a, String key, String value) {
+    protected void set(Annotation a, String key, String value) {
         if (value == null) {
-            return false;
+            return;
         }
         a.addFeature(key, value);
-        return true;
     }
 }
