+++ src/main/groovy/org/lappsgrid/eager/mining/web/nlp/stanford/Worker.groovy
@@ -1,18 +1,16 @@
-package org.lappsgrid.askme.nlp
+package org.lappsgrid.eager.mining.web.nlp.stanford
 
 import com.codahale.metrics.Timer
 import com.codahale.metrics.Meter
-import groovy.transform.TypeChecked
 import groovy.util.logging.Slf4j
-import org.lappsgrid.rabbitmq.Message
-import org.lappsgrid.rabbitmq.topic.PostOffice
+import org.lappsgrid.eager.rabbitmq.Message
+import org.lappsgrid.eager.rabbitmq.topic.PostOffice
 import org.lappsgrid.serialization.DataContainer
 import org.lappsgrid.serialization.Serializer
 
 /**
  *
  */
-@TypeChecked
 @Slf4j("logger")
 class Worker implements Runnable {
 
@@ -20,38 +18,38 @@ class Worker implements Runnable {
     Message message
     PostOffice post
 
-//    Timer timer
-//    Meter counter
-//    Meter errors
+    Timer timer
+    Meter counter
+    Meter errors
 
     Worker(Pipeline pipeline, Message message, PostOffice post, Timer timer, Meter counter, Meter errors) {
         this.pipeline = pipeline
         this.message = message
         this.post = post
-//        this.timer = timer
-//        this.counter = counter
-//        this.errors = errors
+        this.timer = timer
+        this.counter = counter
+        this.errors = errors
         logger.debug("Worker created")
     }
 
     void run() {
         logger.info("Staring worker in thread {}", Thread.currentThread().name)
         DataContainer data
-//        Timer.Context context = timer.time()
+        Timer.Context context = timer.time()
         try {
-            data = Serializer.parse(message.body.toString(), DataContainer)
+            data = Serializer.parse(message.body, DataContainer)
             // TODO Check the discriminator
             data.payload = pipeline.process(data.payload)
-//            counter.mark()
+            counter.mark()
         }
         catch (Exception e) {
             logger.error("Unable to process input.", e)
-//            errors.mark()
+            errors.mark()
             return
         }
-//        finally {
-//            context.close()
-//        }
+        finally {
+            context.close()
+        }
         logger.debug("Sending result to {}", message.route[0])
         message.body = data.asJson()
         post.send(message)
