--- src/main/resources/build
@@ -1,1 +1,0 @@
-36
