+++ src/main/resources/log4j2.properties
@@ -14,12 +14,12 @@ appender.console.layout.type = PatternLayout
 appender.console.layout.pattern = %d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
 
 #Make sure to change log file path as per your need
-property.filename = askme.log
+property.filename = stanford.log
 
 appender.rolling.type = RollingFile
 appender.rolling.name = RollingFile
 appender.rolling.fileName = ${filename}
-appender.rolling.filePattern = askme-%d{yy-MM-dd}.log.gz
+appender.rolling.filePattern = stanford-%d{yy-MM-dd}.log.gz
 appender.rolling.layout.type = PatternLayout
 appender.rolling.layout.pattern = %d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
 appender.rolling.policies.type = Policies
@@ -35,7 +35,7 @@ appender.rolling.strategy.max = 20
 
 #Make sure to change the package structure as per your application
 
-logger.rolling.name = org.lappsgrid.askme.nlp
+logger.rolling.name = org.lappsgrid.eager.web.nlp
 logger.rolling.level = trace
 logger.rolling.additivity = false
 logger.rolling.appenderRef.rolling.ref = RollingFile
