+++ src/main/resources/logback.xml
@@ -22,8 +22,7 @@
 	-->
 
 	<logger name="org.lappsgrid" level="warn"/>
-	<logger name="org.lappsgrid.askme" level="info"/>
-	<logger name="org.lappsgrid.askme.nlp" level="trace"/>
+	<logger name="org.lappsgrid.eager" level="debug"/>
 
 	<root level="debug">
 		<appender-ref ref="STDOUT" />
