+++ src/test/groovy/org/lappsgrid/eager/mining/web/nlp/stanford/Record.groovy
@@ -1,4 +1,4 @@
-package org.lappsgrid.askme.nlp
+package org.lappsgrid.eager.mining.web.nlp.stanford
 /**
  *
  */
